Latest version of the program
To run the application download to server or local machine then navigate
to either 'DecisionSupportSystem\dist' or 'NetBeansProjects\TestData'.
Modify the config.txt files in both directories (one for each
application in the same folder as the *.jar files. Then execute the
*.jar file.

Please note to only enter the folder path in the config folder, do not
put attitional spaces or return characters or it will not work - no
error management in place except for standard IOException alert.